python: relocate python.rb to autoproj/python
